# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Untukmu API",
    "description": "Untukmu Fase 1–5: akun, pesan terenkripsi, sosial, doa, kurasi, pemulihan dan ekspor.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/dashboard/prayers-received": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Prayers Received",
        "operationId": "prayers_received_api_v1_dashboard_prayers_received_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/summary": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Summary",
        "operationId": "summary_api_v1_dashboard_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/dashboard/messages": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "My Messages",
        "operationId": "my_messages_api_v1_dashboard_messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/prayers": {
      "get": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Admin Catalog",
        "operationId": "admin_catalog_api_v1_admin_prayers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/prayers/{tradition}/{prayer_id}": {
      "put": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Curate",
        "operationId": "curate_api_v1_admin_prayers__tradition___prayer_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tradition",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tradition"
            }
          },
          {
            "name": "prayer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prayer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrayerContentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/prayers/{tradition}/{prayer_id}/audio/presign": {
      "post": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Audio Presign",
        "operationId": "audio_presign_api_v1_admin_prayers__tradition___prayer_id__audio_presign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tradition",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tradition"
            }
          },
          {
            "name": "prayer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prayer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrayerAudioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/prayers/{tradition}/{prayer_id}/audio/complete": {
      "post": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Audio Complete",
        "operationId": "audio_complete_api_v1_admin_prayers__tradition___prayer_id__audio_complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tradition",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tradition"
            }
          },
          {
            "name": "prayer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prayer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/prayers/start": {
      "post": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Start",
        "operationId": "start_api_v1_messages__message_id__prayers_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrayerStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{message_id}/prayers": {
      "post": {
        "tags": [
          "Doa dan kurasi"
        ],
        "summary": "Finish",
        "operationId": "finish_api_v1_messages__message_id__prayers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrayerFinish"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrayerAudioInput": {
        "properties": {
          "byte_size": {
            "type": "integer",
            "maximum": 31457280.0,
            "minimum": 100.0,
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "enum": [
              "audio/mpeg",
              "audio/ogg",
              "audio/wav",
              "audio/mp4"
            ],
            "title": "Content Type"
          },
          "license": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "License"
          },
          "attribution": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Attribution"
          },
          "duration_seconds": {
            "type": "integer",
            "maximum": 900.0,
            "minimum": 5.0,
            "title": "Duration Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "byte_size",
          "content_type",
          "license",
          "attribution",
          "duration_seconds"
        ],
        "title": "PrayerAudioInput"
      },
      "PrayerContentInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "translation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation"
          },
          "source_attribution": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Source Attribution"
          },
          "source_url": {
            "type": "string",
            "maxLength": 1000,
            "title": "Source Url",
            "default": ""
          },
          "review_note": {
            "type": "string",
            "maxLength": 1000,
            "title": "Review Note",
            "default": ""
          },
          "reviewed": {
            "type": "boolean",
            "title": "Reviewed",
            "default": false
          },
          "duration_seconds": {
            "type": "integer",
            "maximum": 900.0,
            "minimum": 5.0,
            "title": "Duration Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "source_attribution",
          "duration_seconds"
        ],
        "title": "PrayerContentInput"
      },
      "PrayerFinish": {
        "properties": {
          "playback_token": {
            "type": "string",
            "maxLength": 80,
            "minLength": 40,
            "title": "Playback Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "playback_token"
        ],
        "title": "PrayerFinish"
      },
      "PrayerStart": {
        "properties": {
          "turnstile_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          },
          "catalog_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Catalog Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "catalog_id"
        ],
        "title": "PrayerStart"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}